backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow-cpu>=2.16.0
pillow>=10.0.0
numpy>=1.26.0
python-multipart==0.0.6
pydantic>=2.0.0
